requirements.txt
Flask == 2.1.2
Flask-Limiter == 2.4.5.1
Flask-SQLAlchemy == 2.5.1
PyMySQL == 1.0.2
PyYAML == 6.0